add BINANCE!

+++ src/main/kotlin/bot/telegram/notificator/BalanceInfo.kt
@@ -2,12 +2,13 @@ package bot.telegram.notificator
 
 import bot.telegram.notificator.exchanges.clients.Order
 import bot.telegram.notificator.exchanges.clients.TradePair
+import java.math.BigDecimal
 
 data class BalanceInfo(
     var orderS: Order? = null,
     var orderB: Order? = null,
     var symbols: TradePair,
-    var firstBalance: Double = 0.toDouble(),
-    var secondBalance: Double = 0.toDouble(),
-    var balanceTrade: Double = 0.toDouble()
+    var firstBalance: BigDecimal = 0.toBigDecimal(),
+    var secondBalance: BigDecimal = 0.toBigDecimal(),
+    var balanceTrade: BigDecimal = 0.toBigDecimal()
 )
